+++ extensions/default/src/utils/callInputDialog.tsx
@@ -94,12 +94,12 @@ export async function callInputDialogAutoComplete({
   const dropDownItems = labelConfig ? labelConfig.items : [];
 
   const value = await new Promise<Map<string, string>>((resolve, reject) => {
-    const labellingDoneCallback = value => {
+    const labellingDoneCallback = newValue => {
       uiDialogService.hide('select-annotation');
-      if (typeof value === 'string') {
-        measurement.label = value;
+      if (measurement && typeof newValue === 'string') {
+        measurement.label = newValue;
       }
-      resolve(measurement);
+      resolve(newValue);
     };
 
     uiDialogService.show({