+++ platform/ui-next/src/index.ts
@@ -115,17 +115,9 @@ import {
   WindowLevelHistogram,
 } from './components';
 import { DataRow } from './components/DataRow';
+export * from './components';
+export * from './contextProviders';
 
-import {
-  useNotification,
-  NotificationProvider,
-  useModal,
-  ModalProvider,
-  DialogProvider,
-  useDialog,
-  ManagedDialog,
-} from './contextProviders';
-import { ViewportGridContext, ViewportGridProvider, useViewportGrid } from './contextProviders';
 import * as utils from './utils';
 
 export {
@@ -258,3 +250,4 @@ export {
   WindowLevel,
   WindowLevelHistogram,
 };
+export { utils };